src (1)/src/etl/requirements.txt
psycopg2-binary==2.9.9
shapely==2.0.6
streamlit
pandas
plotly
folium
streamlit-folium
pyproj
